backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-telegram-bot==21.7
python-dotenv==1.0.1
supabase==2.13.0
openai==1.57.0
httpx==0.27.2
pydantic==2.9.2
cryptography==43.0.3
